add csv export
add file patching

+++ src/export/mod.rs
@@ -1,5 +1,5 @@
 use std::fmt::Display;
-use std::fs::File;
+use std::fs::{self, File};
 use std::thread;
 
 use eyre::Result;
@@ -7,6 +7,7 @@ use rfd::FileDialog;
 
 use crate::timing::TimingPoint;
 
+mod csv;
 mod osu;
 
 trait ApplyExportFormat {
@@ -15,12 +16,14 @@ trait ApplyExportFormat {
 
 #[derive(Clone, Copy)]
 pub enum ExportFormat {
+	Csv,
 	Osu,
 }
 
 impl ApplyExportFormat for FileDialog {
 	fn apply_format(self, fmt: ExportFormat) -> Self {
 		match fmt {
+			ExportFormat::Csv => self.add_filter("CSV", &["csv"]),
 			ExportFormat::Osu => self.add_filter("osu! beatmap", &["osu"]),
 		}
 	}
@@ -32,20 +35,29 @@ impl Display for ExportFormat {
 			f,
 			"{}",
 			match self {
-				ExportFormat::Osu => "osu! (.osu)",
+				Self::Csv => "CSV (.csv)",
+				Self::Osu => "osu! (.osu)",
 			}
 		)
 	}
 }
 
 impl ExportFormat {
-	pub fn list() -> &'static [Self] {
+	pub fn game_formats() -> &'static [Self] {
 		&[ExportFormat::Osu]
 	}
 
-	fn run(self, file: File, timing_points: &[TimingPoint]) -> Result<()> {
+	fn create(self, file: File, timing_points: &[TimingPoint]) -> Result<()> {
 		match self {
-			Self::Osu => osu::export(file, timing_points),
+			Self::Csv => csv::create(file, timing_points),
+			Self::Osu => osu::create(file, timing_points),
+		}
+	}
+
+	fn patch(self, file: File, contents: String, timing_points: &[TimingPoint]) -> Result<()> {
+		match self {
+			Self::Csv => csv::patch(file, timing_points),
+			Self::Osu => osu::patch(file, contents, timing_points),
 		}
 	}
 }
@@ -53,9 +65,14 @@ impl ExportFormat {
 pub fn export_timing_points(timing_points: Vec<TimingPoint>, fmt: ExportFormat) {
 	thread::spawn(move || {
 		if let Some(path) = FileDialog::new().apply_format(fmt).save_file() {
-			let file = File::create(path).unwrap();
-
-			let _ = fmt.run(file, &timing_points);
+			if path.exists() {
+				let contents = fs::read_to_string(&path).unwrap();
+				let file = File::create(path).unwrap();
+				let _ = fmt.patch(file, contents, &timing_points);
+			} else {
+				let file = File::create_new(path).unwrap();
+				let _ = fmt.create(file, &timing_points);
+			}
 		}
 	});
 }
